Snapshot 78.0.3904.98

+++ src/content/browser/indexed_db/indexed_db_metadata_coding.cc
@@ -4,6 +4,9 @@
 
 #include "content/browser/indexed_db/indexed_db_metadata_coding.h"
 
+#include <memory>
+#include <utility>
+
 #include "base/strings/string_piece.h"
 #include "content/browser/indexed_db/indexed_db_class_factory.h"
 #include "content/browser/indexed_db/indexed_db_leveldb_coding.h"
@@ -12,7 +15,9 @@
 #include "content/browser/indexed_db/indexed_db_tracing.h"
 #include "content/browser/indexed_db/leveldb/leveldb_env.h"
 #include "content/browser/indexed_db/leveldb/transactional_leveldb_database.h"
+#include "content/browser/indexed_db/leveldb/transactional_leveldb_iterator.h"
 #include "content/browser/indexed_db/leveldb/transactional_leveldb_transaction.h"
+#include "content/browser/indexed_db/scopes/leveldb_scope_deletion_mode.h"
 #include "third_party/blink/public/common/indexeddb/indexeddb_metadata.h"
 
 using base::StringPiece;
@@ -371,7 +376,7 @@ Status ReadDatabaseNamesAndVersionsInternal(
                       database_id, DatabaseMetaDataKey::USER_VERSION),
                   &database_version, &found);
     if (!s.ok() || !found) {
-      INTERNAL_READ_ERROR_UNTESTED(GET_DATABASE_NAMES);
+      INTERNAL_READ_ERROR(GET_DATABASE_NAMES);
       continue;
     }
 
@@ -522,8 +527,8 @@ Status IndexedDBMetadataCoding::CreateDatabase(
     int64_t version,
     IndexedDBDatabaseMetadata* metadata) {
   // TODO(jsbell): Don't persist metadata if open fails. http://crbug.com/395472
-  scoped_refptr<TransactionalLevelDBTransaction> transaction =
-      indexed_db::LevelDBFactory::Get()->CreateLevelDBTransaction(db);
+  std::unique_ptr<LevelDBDirectTransaction> transaction =
+      db->leveldb_class_factory()->CreateLevelDBDirectTransaction(db);
 
   int64_t row_id = 0;
   Status s = indexed_db::GetNewDatabaseId(transaction.get(), &row_id);
@@ -534,16 +539,29 @@ Status IndexedDBMetadataCoding::CreateDatabase(
   if (version == IndexedDBDatabaseMetadata::NO_VERSION)
     version = IndexedDBDatabaseMetadata::DEFAULT_VERSION;
 
-  PutInt(transaction.get(), DatabaseNameKey::Encode(origin_identifier, name),
-         row_id);
-  PutVarInt(
+  s = PutInt(transaction.get(),
+             DatabaseNameKey::Encode(origin_identifier, name), row_id);
+  if (!s.ok()) {
+    INTERNAL_READ_ERROR_UNTESTED(CREATE_IDBDATABASE_METADATA);
+    return s;
+  }
+  s = PutVarInt(
       transaction.get(),
       DatabaseMetaDataKey::Encode(row_id, DatabaseMetaDataKey::USER_VERSION),
       version);
-  PutVarInt(transaction.get(),
-            DatabaseMetaDataKey::Encode(
-                row_id, DatabaseMetaDataKey::BLOB_KEY_GENERATOR_CURRENT_NUMBER),
-            DatabaseMetaDataKey::kBlobKeyGeneratorInitialNumber);
+  if (!s.ok()) {
+    INTERNAL_READ_ERROR_UNTESTED(CREATE_IDBDATABASE_METADATA);
+    return s;
+  }
+  s = PutVarInt(
+      transaction.get(),
+      DatabaseMetaDataKey::Encode(
+          row_id, DatabaseMetaDataKey::BLOB_KEY_GENERATOR_CURRENT_NUMBER),
+      DatabaseMetaDataKey::kBlobKeyGeneratorInitialNumber);
+  if (!s.ok()) {
+    INTERNAL_READ_ERROR_UNTESTED(CREATE_IDBDATABASE_METADATA);
+    return s;
+  }
 
   s = transaction->Commit();
   if (!s.ok()) {
@@ -554,11 +572,12 @@ Status IndexedDBMetadataCoding::CreateDatabase(
   // Note: |version| is not stored on purpose.
   metadata->name = name;
   metadata->id = row_id;
+  metadata->max_object_store_id = 0;
 
   return s;
 }
 
-void IndexedDBMetadataCoding::SetDatabaseVersion(
+leveldb::Status IndexedDBMetadataCoding::SetDatabaseVersion(
     TransactionalLevelDBTransaction* transaction,
     int64_t row_id,
     int64_t version,
@@ -567,7 +586,7 @@ void IndexedDBMetadataCoding::SetDatabaseVersion(
     version = IndexedDBDatabaseMetadata::DEFAULT_VERSION;
   DCHECK_GE(version, 0) << "version was " << version;
   c->version = version;
-  PutVarInt(
+  return PutVarInt(
       transaction,
       DatabaseMetaDataKey::Encode(row_id, DatabaseMetaDataKey::USER_VERSION),
       version);
@@ -604,7 +623,7 @@ Status IndexedDBMetadataCoding::CreateObjectStore(
   if (!s.ok())
     return s;
 
-  static const constexpr long long kInitialLastVersionNumber = 1;
+  static const constexpr int64_t kInitialLastVersionNumber = 1;
   const std::string name_key = ObjectStoreMetaDataKey::Encode(
       database_id, object_store_id, ObjectStoreMetaDataKey::NAME);
   const std::string key_path_key = ObjectStoreMetaDataKey::Encode(
@@ -625,16 +644,34 @@ Status IndexedDBMetadataCoding::CreateObjectStore(
           ObjectStoreMetaDataKey::KEY_GENERATOR_CURRENT_NUMBER);
   const std::string names_key = ObjectStoreNamesKey::Encode(database_id, name);
 
-  PutString(transaction, name_key, name);
-  PutIDBKeyPath(transaction, key_path_key, key_path);
-  PutInt(transaction, auto_increment_key, auto_increment);
-  PutInt(transaction, evictable_key, false);
-  PutInt(transaction, last_version_key, kInitialLastVersionNumber);
-  PutInt(transaction, max_index_id_key, kMinimumIndexId);
-  PutBool(transaction, has_key_path_key, !key_path.IsNull());
-  PutInt(transaction, key_generator_current_number_key,
-         ObjectStoreMetaDataKey::kKeyGeneratorInitialNumber);
-  PutInt(transaction, names_key, object_store_id);
+  s = PutString(transaction, name_key, name);
+  if (!s.ok())
+    return s;
+  s = PutIDBKeyPath(transaction, key_path_key, key_path);
+  if (!s.ok())
+    return s;
+  s = PutInt(transaction, auto_increment_key, auto_increment);
+  if (!s.ok())
+    return s;
+  s = PutInt(transaction, evictable_key, false);
+  if (!s.ok())
+    return s;
+  s = PutInt(transaction, last_version_key, kInitialLastVersionNumber);
+  if (!s.ok())
+    return s;
+  s = PutInt(transaction, max_index_id_key, kMinimumIndexId);
+  if (!s.ok())
+    return s;
+  s = PutBool(transaction, has_key_path_key, !key_path.IsNull());
+  if (!s.ok())
+    return s;
+  s = PutInt(transaction, key_generator_current_number_key,
+             ObjectStoreMetaDataKey::kKeyGeneratorInitialNumber);
+  if (!s.ok())
+    return s;
+  s = PutInt(transaction, names_key, object_store_id);
+  if (!s.ok())
+    return s;
 
   metadata->name = std::move(name);
   metadata->id = object_store_id;
@@ -670,21 +707,28 @@ Status IndexedDBMetadataCoding::DeleteObjectStore(
 
   s = transaction->RemoveRange(
       ObjectStoreMetaDataKey::Encode(database_id, object_store.id, 0),
-      ObjectStoreMetaDataKey::EncodeMaxKey(database_id, object_store.id), true);
+      ObjectStoreMetaDataKey::EncodeMaxKey(database_id, object_store.id),
+      LevelDBScopeDeletionMode::kImmediateWithRangeEndInclusive);
 
   if (s.ok()) {
-    transaction->Remove(
+    s = transaction->Remove(
         ObjectStoreNamesKey::Encode(database_id, object_store_name));
+    if (!s.ok()) {
+      INTERNAL_WRITE_ERROR_UNTESTED(DELETE_OBJECT_STORE);
+      return s;
+    }
 
     s = transaction->RemoveRange(
         IndexFreeListKey::Encode(database_id, object_store.id, 0),
-        IndexFreeListKey::EncodeMaxKey(database_id, object_store.id), true);
+        IndexFreeListKey::EncodeMaxKey(database_id, object_store.id),
+        LevelDBScopeDeletionMode::kImmediateWithRangeEndInclusive);
   }
 
   if (s.ok()) {
     s = transaction->RemoveRange(
         IndexMetaDataKey::Encode(database_id, object_store.id, 0, 0),
-        IndexMetaDataKey::EncodeMaxKey(database_id, object_store.id), true);
+        IndexMetaDataKey::EncodeMaxKey(database_id, object_store.id),
+        LevelDBScopeDeletionMode::kImmediateWithRangeEndInclusive);
   }
 
   if (!s.ok())
@@ -721,9 +765,21 @@ Status IndexedDBMetadataCoding::RenameObjectStore(
   const std::string old_names_key =
       ObjectStoreNamesKey::Encode(database_id, metadata->name);
 
-  PutString(transaction, name_key, new_name);
-  PutInt(transaction, new_names_key, metadata->id);
-  transaction->Remove(old_names_key);
+  s = PutString(transaction, name_key, new_name);
+  if (!s.ok()) {
+    INTERNAL_READ_ERROR_UNTESTED(DELETE_OBJECT_STORE);
+    return s;
+  }
+  s = PutInt(transaction, new_names_key, metadata->id);
+  if (!s.ok()) {
+    INTERNAL_READ_ERROR_UNTESTED(DELETE_OBJECT_STORE);
+    return s;
+  }
+  s = transaction->Remove(old_names_key);
+  if (!s.ok()) {
+    INTERNAL_READ_ERROR_UNTESTED(DELETE_OBJECT_STORE);
+    return s;
+  }
   *old_name = std::move(metadata->name);
   metadata->name = std::move(new_name);
   return s;
@@ -756,10 +812,18 @@ Status IndexedDBMetadataCoding::CreateIndex(
   const std::string multi_entry_key = IndexMetaDataKey::Encode(
       database_id, object_store_id, index_id, IndexMetaDataKey::MULTI_ENTRY);
 
-  PutString(transaction, name_key, name);
-  PutBool(transaction, unique_key, is_unique);
-  PutIDBKeyPath(transaction, key_path_key, key_path);
-  PutBool(transaction, multi_entry_key, is_multi_entry);
+  s = PutString(transaction, name_key, name);
+  if (!s.ok())
+    return s;
+  s = PutBool(transaction, unique_key, is_unique);
+  if (!s.ok())
+    return s;
+  s = PutIDBKeyPath(transaction, key_path_key, key_path);
+  if (!s.ok())
+    return s;
+  s = PutBool(transaction, multi_entry_key, is_multi_entry);
+  if (!s.ok())
+    return s;
 
   metadata->name = std::move(name);
   metadata->id = index_id;
@@ -782,8 +846,9 @@ Status IndexedDBMetadataCoding::DeleteIndex(
       IndexMetaDataKey::Encode(database_id, object_store_id, metadata.id, 0);
   const std::string index_meta_data_end =
       IndexMetaDataKey::EncodeMaxKey(database_id, object_store_id, metadata.id);
-  Status s = transaction->RemoveRange(index_meta_data_start,
-                                      index_meta_data_end, true);
+  Status s = transaction->RemoveRange(
+      index_meta_data_start, index_meta_data_end,
+      LevelDBScopeDeletionMode::kImmediateWithRangeEndExclusive);
   return s;
 }
 
@@ -801,8 +866,9 @@ Status IndexedDBMetadataCoding::RenameIndex(
       database_id, object_store_id, metadata->id, IndexMetaDataKey::NAME);
 
   // TODO(dmurph): Add consistancy checks & umas for old name.
-
-  PutString(transaction, name_key, new_name);
+  leveldb::Status s = PutString(transaction, name_key, new_name);
+  if (!s.ok())
+    return s;
   *old_name = std::move(metadata->name);
   metadata->name = std::move(new_name);
   return Status::OK();
